agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.12.0",
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "langchain>=1.0.8",
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=1.0.3",
    "langchain-community>=0.3.0",
    "langgraph>=1.0.3",
    "openai>=2.8.1",
    "pdf2image>=1.17.0",
    "pillow>=12.0.0",
    "protobuf>=4.25.0",
    "pymupdf>=1.26.6",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.9",
    "qwen-vl-utils>=0.0.14",
    "sentencepiece>=0.2.0",
    "tavily-python>=0.5.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "transformers>=4.57.1",
    "uvicorn>=0.32.0",
    "pinecone>=8.1.1",
]
